Better categories migration using database method

+++ trunk/admin/includes/jupgrade.category.class.php
@@ -49,20 +49,13 @@ class JUpgradeproCategory extends JUpgradepro
 	 */
 	public function beforeHook()
 	{
-		// Insert uncategorized id
 		$query = $this->_db->getQuery(true);
-		$query->insert('#__jupgradepro_categories')->columns('`old`, `new`')->values("0, 2");
-		try {
-			$this->_db->setQuery($query)->execute();
-		} catch (RuntimeException $e) {
-			throw new RuntimeException($e->getMessage());
-		}
 
 		if ($this->params->keep_ids == 1)
 		{
 			// Getting the categories
 			$query->clear();
-			$query->select("`id`, `parent_id`, `path`, `extension`, `title`, `alias`, `note`, `description`, `published`,  `params`, `created_user_id`");
+			$query->select("`id`, `parent_id`, `path`, `extension`, `title`, `alias`, `note`, `description`, `published`, `checked_out`, `checked_out_time`, `access`, `metadesc`, `metakey`, `metadata`, `params`, `created_user_id`, `modified_user_id`, `modified_time`, `hits`, `language`, `version`");
 			$query->from("#__categories");
 			$query->where("id > 1");
 			$query->order('id ASC');
@@ -81,49 +74,18 @@ class JUpgradeproCategory extends JUpgradepro
 
 				$this->_db->insertObject('#__jupgradepro_default_categories', $category);
 
-				// Getting the categories table
-				$table = JTable::getInstance('Category', 'JTable');
-				// Load it before delete. Joomla bug?
-				$table->load($id);
-				// Delete
-				$table->delete($id);
+				if (get_class($this) == "JUpgradeproCategories")
+				{
+					// Getting the categories table
+					$table = JTable::getInstance('Category', 'JTable');
+					// Load it before delete. Joomla bug?
+					$table->load($id);
+					// Delete
+					$table->delete($id);
+				}
 			}
 		}
 	}
-
-	/**
-	 * Get the raw data for this part of the upgrade.
-	 *
-	 * @return	array	Returns a reference to the source data array.
-	 * @since	0.5.6
-	 * @throws	Exception
-	 */
-	public function databaseHook($rows = null)
-	{
-		// Do some custom post processing on the list.
-		foreach ($rows as &$row)
-		{
-			$row = (array) $row;
-
-			$row['params'] = $this->convertParams($row['params']);
-			$row['title'] = str_replace("'", "&#39;", $row['title']);
-			$row['description'] = str_replace("'", "&#39;", $row['description']);
-
-			if ($row['extension'] == 'com_banner') {
-				$row['extension'] = "com_banners";
-			}else if ($row['extension'] == 'com_contact_details') {
-				$row['extension'] = "com_contact";
-			}
-
-			// Correct alias
-			if ($row['alias'] == "") {
-				$row['alias'] = JFilterOutput::stringURLSafe($row['title']);
-			}
-		}
-
-		return $rows;
-	}
-
 
 	/**
 	 * Sets the data in the destination database.
@@ -177,7 +139,7 @@ class JUpgradeproCategory extends JUpgradepro
 	 */
 	public function insertCategory($row, $parent = false)
 	{
-		// Getting the category table
+		// Get the category table
 		$category = JTable::getInstance('Category', 'JTable', array('dbo' => $this->_db));
 
 		// Disable observers calls
@@ -198,42 +160,60 @@ class JUpgradeproCategory extends JUpgradepro
 		$rules['core.create'] = $rules['core.delete'] = $rules['core.edit'] = $rules['core.edit.state'] = $rules['core.edit.own'] = '';
 		$row['rules'] = $rules;
 
-		// Correct extension
-		if (isset($row['extension'])) {
-			if ($this->params->keep_ids == 1 && is_numeric($row['extension']) || $row['extension'] == "" || $row['extension'] == "category") {
-				$row['extension'] = "com_content";
-			}
-
-			// Fixing extension name if it's section
-			if ($row['extension'] == 'com_section') {
-				$row['id'] = 0;
-				$row['extension'] = "com_content";
-				$category->setLocation(1, 'last-child');
-			}
-		}
-
 		// Fix language
 		$row['language'] = !empty($row['language']) ? $row['language'] : '*';
 
 		// Check if path is correct
 		$row['path'] = empty($row['path']) ? $row['alias'] : $row['path'];
 
+		// Get alias from title if its empty
+		if ($row['alias'] == "") {
+			$row['alias'] = JFilterOutput::stringURLSafe($row['title']);
+		}
+
 		// Check if has duplicated aliases
-		$alias = $this->getAlias('#__categories', $row['alias']);
+		$alias = $this->getAlias('#__categories', $row['alias'], $row['extension']);
 
 		// Prevent MySQL duplicate error
 		// @@ Duplicate entry for key 'idx_client_id_parent_id_alias_language'
 		$row['alias'] = (!empty($alias)) ? $alias."~" : $row['alias'];
 
 		// Remove the default id if keep ids parameters is not enabled
-		if ($this->params->keep_ids != 1) {
+		if ($this->params->keep_ids != 1)
+		{
+			// Unset id
 			unset($row['id']);
 
-			if ($row['parent_id'] != 1) {
+			// Save the parent if old installation is 2.5 or greater
+			if (version_compare(JUpgradeproHelper::getVersion('old'), '1.5', '>') && $row['parent_id'] != 1)
+			{
 				$oldlist->section = $row['parent_id'];
 			}else{
 				$parent = 1;
 			}
+		}else if ($this->params->keep_ids == 1){
+
+			// Save section id if old Joomla! version is 1.0
+			if (version_compare(JUpgradeproHelper::getVersion('old'), '1.0', '=') && isset($row['section']))
+			{
+				$oldlist->section = $row['section'];
+			}else	if (version_compare(JUpgradeproHelper::getVersion('old'), '1.5', '>') && $row['parent_id'] != 1) {
+				$oldlist->section = $row['parent_id'];
+			}
+		}
+
+		// Correct extension
+		if (isset($row['extension'])) {
+			if (is_numeric($row['extension']) || $row['extension'] == "" || $row['extension'] == "category") {
+				$row['extension'] = "com_content";
+			}
+		}
+
+		// Fixing extension name if it's section
+		if ($row['extension'] == 'com_section') {
+			unset($row['id']);
+			$row['extension'] = "com_content";
+			$parent = 1;
 		}
 
 		// If has parent made $path and get parent id
@@ -261,5 +241,96 @@ class JUpgradeproCategory extends JUpgradepro
 		}
 
 	 	return true;
+	}
+
+	/**
+	 * Insert existing categories saved in cleanup step
+	 *
+	 * @return	void
+	 * @since	3.0
+	 */
+	protected function insertExisting()
+	{
+		// Getting the database instance
+		$db = JFactory::getDbo();
+
+		// Getting the data
+		$query = $db->getQuery(true);
+		$query->select('*');
+		$query->from('#__jupgradepro_default_categories');
+		$query->order('id ASC');
+		$db->setQuery($query);
+		$categories = $db->loadAssocList();
+
+		foreach ($categories as $category) {
+			// Unset id
+			$category['id'] = 0;
+
+			// Looking for parent
+			$parent = 1;
+			$explode = explode("/", $category['path']);
+
+			if (count($explode) > 1) {
+
+				// Getting the data
+				$query = $db->getQuery(true);
+				$query->select('id');
+				$query->from('#__categories');
+				$query->where("path = '{$explode[0]}'");
+				$query->order('id ASC');
+				$query->limit(1);
+
+				$db->setQuery($query);
+				$parent = $db->loadResult();
+			}
+
+			// Inserting the category
+			$this->insertCategory($category, $parent);
+		}
+
+	}
+
+	/**
+	 * Return the root id.
+	 *
+	 * @return	int	The last root id
+	 * @since		3.2.2
+	 * @throws	Exception
+	 */
+	public function getRootNextId()
+	{
+		$query = $this->_db->getQuery(true);
+		$query->select("`id` + 1");
+		$query->from("#__categories");
+		$query->where("id > 1");
+		$query->order('id DESC');
+		$query->limit(1);
+		$this->_db->setQuery($query);
+
+		return (int) $this->_db->loadResult();
+	}
+
+	/**
+	 * Get the saved first category
+	 *
+	 * @return object An object with the first category.
+	 * @since		3.3.0
+	 * @throws	Exception
+	 */
+	public function getFirstCategory()
+	{
+		// Get the correct equipment
+		$query = $this->_db->getQuery(true);
+		// Select some values
+		$query->select("*");
+		// Set the from table
+		$query->from($this->_db->qn('#__jupgradepro_default_categories') . ' as c');
+		// Conditions
+		$query->where("c.root_id = 1");
+		// Limit and order
+		$query->order('id DESC');
+		$query->limit(1);
+		// Retrieve the data.
+		return $this->_db->setQuery($query)->loadAssoc();
 	}
 }
